Add padding to book-content, use calc to breakout
Previously, the main book-content container didn't have any
padding.  Certain elements would specify a margin, but then
there were issues with nested elements having it applied twice.

This removes all the margins and sets padding on the main
container.  The `book-content-full-width` mixin uses calc
to set the width to '100% + padding*2' in order to fill the
container.