Add options prop to loadable-item for passing to store
With this property options can be passed to the store.  It's
intended to be used for stores that need to pass multiple
parameters to their endpoints.

For instance the learning guide for a teacher viewing a student
must pass the courseId and roleId to the endpoint when making a request.