Create route for marking task as complete
When a link is right or middle mouse button clicked to open
in a new window, the current page does not receive any events.  This
means the link will be loaded, but the step not marked as completed
since the onClick handler is never called.

To work around this, we setup a new destination url and have the link
point to that.  If the onClick hander is called, we open the external
link in a new window (tab) and mark the step as complete.  If the
handler isn't called and the destination page is loaded, we mark it
complete there and open the window at that time.