Scroll more if end position isn't correct
If the document is in the process of being manipulated (adding
exercises, resizing images, etc) when scrolling is initiated, the end
position may not be correct.  In that case mimic the browser behaviour
and scroll some more until it is correct.  A failsafe 3 time limit is
set to ensure that it doesn't continue to jump around infinitely