Render even if course isn't (yet) loaded
Some of the specs and most likely production code use the browse the
book component to link to a book for a course who's flux store isn't yet
loaded.

We go ahead and render on the assumption that the course ID is valid and the
reference book will be able to load it once it's instantiated.