Move logic into single AccountsIframe component
After logging out, accounts will now display the login page itself,
but won't convey that information directly through the iframe.

As a work-around, we sniff the current page through the page load
events. If a user is logged out, but we notice that the profile page is
loaded we start a fresh login process.  That will kick accounts to
notify us about the user's current login information