Downgrade lodash to 3.10
Upgrading to version 4.x breaks the calendar.  It's getEventHandlers
uses _.pick with a funtion on props to select the valid props.

Lodash 4.x has removed the ability to use a function with _.pick, and
requires the use of _.pickBy for that.

Because of this, the onClick prop isn't passed to the Day component and
days cannot be clicked.