Expose bookUUID and project to React context
While I'm not a huge fan of using context (smells like a global var), in this case I think it makes sense.

The exercise trouble link needs to know what books is being displayed but it's really deeply nested (12 levels by my
count), it would be painful to pass these vars through each level, and we'd probably miss a few code branches.

A better solution might be to use a broadcast type approach but that also seems overkill for now