Set data-section on wrapper element and increment it on parts
This fixes two bugs, the first is that if the stimulus has a large element in it (like a video), it would stretch quite
a bit above the data-section marker, and wouldn't match the scroll position.

The other is that the inner parts of the exercise all had the same step-index as the exercise.  We fix that by not
setting one if it's the first, and then incrementing one for the other parts