use action so setting highlighter instance waits to re-render
Since highlighter is observable, it causes a re-render when it's set.

Which is what we want, except the code was sometimes re-rendering before the
annotations were attached.

To prevent that the function is ran inside a "action", which makes
mobx wait until the function completes before it fires updates.

The "once" wrapping wasn't having any affect since the function was
re-created ever time