fix course timezone not being used when loading and saving assignment
This fixes a regression with the course timezone not being used with
moment-timezone when sending the field value to the DateTimeInput
component. With this fix, the value loaded should always match the
course's timezone. This also changes dataForSave to use the Luxon
methods, and fixes a bug with it not converting to the course's
timezone before formatting with tokens. This is necessary because
the backend doesn't take TZ info from the front end, so the
formatted string needs to match the course TZ.

Misc:

- Convert plan details to use Time/Luxon methods to make datetime
  TZ match the other screens
- moment-timezone is updated to get type info for Typescript
- use data-test-ids for spec, remove TZ replace now that plan
  details displays them