fix course timezone not being used when loading and saving assignment (#3771)
* Re-enabled e2e assignment tests and check relative assignment dates

* Fix type error on $eval

* Ensure the resulting due and close dates match the input fields

* Added failing assignment edit spec with browser in different timezone

* Fixed e2e lint

* Use data-test-id instead of classes for the tests
Extracted common selectors to the top of the assignment-review e2e file
Increased playwright timeout (needed for CI)

* Increased the playwright timeout back to 60 seconds

* fix course timezone not being used when loading and saving assignment

This fixes a regression with the course timezone not being used with
moment-timezone when sending the field value to the DateTimeInput
component. With this fix, the value loaded should always match the
course's timezone. This also changes dataForSave to use the Luxon
methods, and fixes a bug with it not converting to the course's
timezone before formatting with tokens. This is necessary because
the backend doesn't take TZ info from the front end, so the
formatted string needs to match the course TZ.

Misc:

- Convert plan details to use Time/Luxon methods to make datetime
  TZ match the other screens
- moment-timezone is updated to get type info for Typescript
- use data-test-ids for spec, remove TZ replace now that plan
  details displays them

* test waiting for modal to appear before clicking

* wip

* improve tests for DateTimeInput

* wip time zone shenanigans

* update spec and add comments

* fix issue with formatting timezone after date-input-changes

* fix lint

* actually fix lint

* remove debug field

* test removing timeout

* improve flakiness of spec

* try a very high number

* Replaced custom wait in visitPage with waiting until domcontentloaded
Removed custom timeouts everywhere

* Use networkidle instead of domcontentloaded
Ensure we set the playwright wait action timeouts

* Set the defaultTimeouts during global setup on the page, not the browser

* Do navigation and waiting in 1 command, increase timeouts

Co-authored-by: Dante Soares <dante.m.soares@rice.edu>
Co-authored-by: tom <tom.g.woodward@gmail.com>